fix data type total

+++ src/main/java/com/luis/market/persistence/entity/ComprasProducto.java
@@ -35,11 +35,11 @@ public class ComprasProducto {
         this.cantidad = cantidad;
     }
 
-    public Double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
